+++ embrace_ios/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.5.0
+
+* Added getCurrentSessionId() method to get the ID of the current session.
+* Renamed some functions to improve consistency between the Embrace SDKs (old names have been deprecated).
+* Updated Embrace Android SDK to 5.24.0
+* Updated Embrace iOS SDK to 5.23.1
+
 # 1.4.0
 
 * Added logHandledDartError() method to record exceptions that have been handled.